refactor(config): promote deployment-config models to agentex.config.*
Promotes the deployment/agent configuration model classes out of the heavy
ADK (agentex.lib.sdk.config) into a new slim-shipped, hand-authored namespace
agentex.config.*, so REST-only consumers can import them from the slim
agentex-client wheel without installing the ADK runtime. Deployment-models
analog of #371 (which did this for protocol types).

- New agentex.config.{agent_config,build_config,deployment_config,
  local_development_config,environment_config,agent_manifest} plus model deps
  {credentials,agent_configs}. ConfigBaseModel inlines the ConfigDict the
  former model_utils.BaseModel provided (from_attributes/populate_by_name),
  folding in DeploymentConfig/InjectedSecretsValues' nested validate_by_name.
- environment_config: model classes promoted; from_yaml/yaml loading stays in
  lib as the load_environments_config free function (keeps models slim-safe).
- agent_manifest: AgentManifest model promoted (egp-api-backend validates
  manifest dicts with it); from_yaml -> load_agent_manifest, context_manager ->
  build_context_manager, plus BuildContextManager stay in lib.
- Back-compat shims left at the old agentex.lib.sdk.config.* and
  agentex.lib.types.{credentials,agent_configs} paths.
- Internal lib consumers + tests repointed to the canonical paths.
- Modernize the promoted agent_configs/local_development_config validators to
  @field_validator so the new public surface emits no pydantic-v1 deprecation.
