+++ src/agentex/lib/cli/utils/path_utils.py
@@ -4,7 +4,7 @@ from typing import Dict
 from pathlib import Path
 
 from agentex.lib.utils.logging import make_logger
-from agentex.lib.sdk.config.agent_manifest import AgentManifest
+from agentex.config.agent_manifest import AgentManifest
 
 logger = make_logger(__name__)
